[ll] Move SwapChain into window for GL presentation
